Move CreateTable rendering to SqlRenderer
This will let us disentangle rendering, and reuse the render method
in a future RedefineTable step without messing with existing steps on
sqlite.
